Introducing v2.0
Back story:
Chrome made a change such that the old new tab page could not be accessed. (#1) So we update the app to not override the new tab page, but instead style the default one so that it looks more or less the same. We lose the date and time but gain the bookmarks on the top. Fair trade. It was a good ride while it lasted. :)